isolated example where minting floating point errors add up

+++ mechafil_jax/minting.py
@@ -11,7 +11,7 @@ from functools import partial
 from .constants import EXA, EXBI, PIB, NETWORK_START
 from .date_utils import datetime64_delta_to_days
 
-LAMBDA = np.log(2) / (
+LAMBDA = jnp.log(2) / (
     6.0 * 365
 )  # minting exponential reward decay rate (6yrs half-life)
 FIL_BASE = 2_000_000_000.0
@@ -53,7 +53,7 @@ def compute_minting_trajectory_df(
     capped_power_reference = 'network_RBP_EIB' if minting_base == 'rbp' else 'network_QAP_EIB'
 
     minting_dict = {
-        "days": jnp.arange(start_day, end_day),
+        "days": jnp.arange(start_day, end_day).astype(jnp.float32),
         "network_RBP_EIB": rb_total_power_eib,
         "network_QAP_EIB": qa_total_power_eib,
         "day_onboarded_power_QAP_PIB": qa_day_onboarded_power_pib,
@@ -73,13 +73,15 @@ def compute_minting_trajectory_df(
     # Add cumulative rewards and get daily rewards minted
     minting_dict["cum_network_reward"] = minting_dict["cum_baseline_reward"] + minting_dict["cum_simple_reward"]
     cum_network_reward_zero = minting_dict["cum_network_reward"][0]
+    # print(cum_network_reward_zero)
     
-    day_network_reward = jnp.zeros(len(minting_dict["cum_network_reward"])+1).astype(jnp.float32)
+    day_network_reward = jnp.zeros(len(minting_dict["cum_network_reward"])+1)
     day_network_reward = day_network_reward.at[1:].set(minting_dict["cum_network_reward"])
     day_network_reward = day_network_reward.at[0].set(cum_network_reward_zero) # equiv. to prepend in NP
     day_network_reward = jnp.diff(day_network_reward)
     day_network_reward = day_network_reward.at[0].set(day_network_reward[1]) # to match mechaFIL
     minting_dict["day_network_reward"] = day_network_reward
+    # print(day_network_reward.dtype)
 
     return minting_dict
